Restructured the get content scripts
