+++ src/Ecandata/well_data/H38-0004.csv
@@ -1,0 +1,31 @@
+Site ID,H38/0004
+Site Name,H38/0004
+Sample ID,20242086
+Collection Date,06-Mar 2024 07:00am
+Name,Value,Units
+"Alkalinity, Total","15.2","g/m3 as CaCO3"
+"Ammoniacal Nitrogen","<0.010","mg/L"
+"Bicarbonate","18.500","mg/L  at 25°C"
+"Calcium, Dissolved","3.000","mg/L"
+"Chloride","0.600","mg/L"
+"Conductivity","2.9","mS/m"
+"Conductivity (Field)","2.98","mS/m"
+"Difference in Ion Balance","3.000","%"
+"Dissolved Oxygen","8.600","mg/L"
+"Dissolved Oxygen Saturation","81.000","%"
+"Dissolved Reactive Phosphorus","0.0053","mg/L"
+"E. coli","<1","MPN/100mL"
+"Hardness, Total","9.800","g/m3 as CaCO3"
+"Magnesium, Dissolved","0.560","mg/L"
+"Nitrate Nitrogen","0.090","mg/L"
+"Oxidation Reduction Potential","103.200","mV"
+"pH","6.6",""
+"pH (Field)","6.83",""
+"Potassium, Dissolved","0.540","mg/L"
+"Silica, Reactive","9.700","mg/L  as SiO2"
+"Sodium, Dissolved","2.500","mg/L"
+"Sulphate","0.700","mg/L"
+"Sum of anions","0.340","meq/L"
+"Sum of cations","0.320","meq/L"
+"Total Coliforms","<1.00","number/100mL"
+"Water Temperature (Field)","10.5","C"
