+++ src/Ecandata/well_data/I39-0007.csv
@@ -1,0 +1,31 @@
+Site ID,I39/0007
+Site Name,I39/0007
+Sample ID,20242125
+Collection Date,04-Mar 2024 11:51am
+Name,Value,Units
+"Alkalinity, Total","51.0","g/m3 as CaCO3"
+"Ammoniacal Nitrogen","<0.010","mg/L"
+"Bicarbonate","62.000","mg/L  at 25°C"
+"Calcium, Dissolved","16.900","mg/L"
+"Chloride","2.300","mg/L"
+"Conductivity","13.3","mS/m"
+"Conductivity (Field)","13.34","mS/m"
+"Difference in Ion Balance","1.540","%"
+"Dissolved Oxygen","6.310","mg/L"
+"Dissolved Oxygen Saturation","66.300","%"
+"Dissolved Reactive Phosphorus","0.0121","mg/L"
+"E. coli","<1","MPN/100mL"
+"Hardness, Total","51.000","g/m3 as CaCO3"
+"Magnesium, Dissolved","2.000","mg/L"
+"Nitrate Nitrogen","1.580","mg/L"
+"Oxidation Reduction Potential","199.400","mV"
+"pH","7.4",""
+"pH (Field)","6.46",""
+"Potassium, Dissolved","1.530","mg/L"
+"Silica, Reactive","14.500","mg/L  as SiO2"
+"Sodium, Dissolved","7.900","mg/L"
+"Sulphate","7.600","mg/L"
+"Sum of anions","1.350","meq/L"
+"Sum of cations","1.390","meq/L"
+"Total Coliforms","<1.00","number/100mL"
+"Water Temperature (Field)","14.7","C"
